Add modified version of sanfran_urbansim as test model
This commit involves, regrettably, a large number of hacky
changes that represent a mashup of sanfran_urbansim
and urbansim_defaults. This goal is to have a good replacement
of the two that has data ready to go and a reasonable configuration,
while maintaining some generalizability.

The current version of this sf_example directory is tenuous
and we'll need to do some serious work. This commit was to bring it up
to compatibility with the new pro forma module.